requirements.txt
feedparser
openai
requests
beautifulsoup4
PyYAML
